Remove runtime error on form close
Added the global variable 'runProgram' to ensure that the user didn't close the user form. This variable's value is only set to TRUE if the user clicks RUN on the user form. If they close the form, this variable's value remains FALSE and the program closes when the 'If Not (runProgram)' test is performed.